Move API utils from core to api app and remove dead code
- Move get_count_safely and build_error_response from apps/core/utils.py
  to apps/api/utils.py
- Remove duplicate functions from apps/tasks/utils.py
- Remove unused get_related_object_safely from apps/tasks/utils.py
- Remove unused test utilities (format_task_data, get_system_uuid,
  get_related_object_safely) from tests/test_utils.py
- Move and rename test files from tests/unit/core/test_core_utils*.py
  to tests/unit/api/test_api_utils*.py
- Remove tests for dead code (~38 tests removed)
- Delete apps/core/utils.py (now empty)
