+++ apps/api/v1/base_serializers.py
@@ -6,7 +6,7 @@ from typing import Any
 
 from rest_framework import serializers
 
-from apps.core.utils import get_count_safely
+from apps.api.utils import get_count_safely
 
 
 class CountFieldMixin:
