+++ apps/api/v1/tasks/views.py
@@ -49,8 +49,8 @@ from rest_framework import status
 from rest_framework.decorators import action
 from rest_framework.response import Response
 
+from apps.api.utils import build_error_response
 from apps.core.permissions import DeveloperModeRequired
-from apps.core.utils import build_error_response
 from apps.tasks.models import Task, TaskExecution
 
 from ..base_views import BaseViewSet
