+++ tests/unit/core/test_mixins_comprehensive.py
@@ -10,7 +10,6 @@ from django.utils import timezone
 
 from apps.tasks.mixins import StatusTrackingMixin, TimestampMixin
 
-
 # ============================================================================
 # TimestampMixin Tests
 # ============================================================================
